refactor(viewer): extract viewer_alerts.jsx from viewer.jsx
Pure file split, no behavior change. Step 3 of mp-pxxc.
